requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
onnxruntime==1.20.1
numpy==1.26.4
python-multipart==0.0.18
Pillow==11.0.0
openai==1.58.1
pydantic==2.10.3
python-dotenv==1.0.1
